warnings (x64)

+++ scilab/modules/graphics/src/c/getHandleProperty/get_links_property.c
@@ -40,7 +40,7 @@ int get_links_property( sciPointObj * pobj )
     return -1 ;
   }
   for (i=0; i<nblegends; i++) {
-    handles[i]=pLEGEND_FEATURE(pobj)->tabofhandles[i];
+    handles[i]=(long)pLEGEND_FEATURE(pobj)->tabofhandles[i];
   }
   status = sciReturnRowHandleVector(handles, nblegends);
